% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col_df.R
\name{add_col_df}
\alias{add_col_df}
\title{Add a column with value to dataframes}
\usage{
add_col_df(df, col, col_val = NULL)
}
\arguments{
\item{df}{Dataframes}

\item{col}{Label for column name}

\item{col_val}{value for column}
}
\value{
list of dataframes
}
\description{
Add a column with value to dataframes
}
