% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col_df_list.R
\name{add_col_df_list}
\alias{add_col_df_list}
\title{Add a column with value to all dataframes in a list}
\usage{
add_col_df_list(df_list, col, col_val)
}
\arguments{
\item{df_list}{list of dataframes}

\item{col}{Label for column name}

\item{col_val}{value for column}
}
\value{
list of dataframes
}
\description{
Add a column with value to all dataframes in a list
}
