% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nclust_cols.R
\name{add_nclust_cols}
\alias{add_nclust_cols}
\title{Add "np" columns to a dataframe.}
\usage{
add_nclust_cols(df, nr)
}
\arguments{
\item{df}{the dataframe to add columns to}

\item{np}{the number of columns to add.}
}
\value{
df
}
\description{
Representing the clusters a column for the clust distance
  is added for each cluster.
  The label for each column is given by the function [make_clust_cols].
}
\seealso{
Other dataframe_editors: 
\code{\link{add_np_cols}()}
}
\concept{dataframe_editors}
