% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_cent.R
\name{calc_clust_cent}
\alias{calc_clust_cent}
\title{Check convergence for the cluster centroid}
\usage{
calc_clust_cent(clustnum_df, b_mat, p_mat, bin_p_clust = TRUE)
}
\arguments{
\item{clustnum_df}{The cluster membership dataframe. The rows are the snps,
The columns are "clust_num".}

\item{b_mat}{The data for each snp. Each row is the snps coordinate.}

\item{p_mat}{Matrix of p-values corresponding to b_mat}

\item{bin_p_clust}{Bool switch 1 if pvalues should be used to weight averages
of snps}
}
\value{
out_centroid_df
}
\description{
Check convergence for the cluster centroid
}
\details{
If the cluster is not empty then the new centres are found using
    the means of the snps in the cluster.
  If the bin_p_clust then weight the average by the p-values.
}
\seealso{
Other clustering_components: 
\code{\link{calc_clust_prob}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{check_clust_cent}()},
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{get_aic}()},
\code{\link{km_nan}()},
\code{\link{make_clust_col_name}()},
\code{\link{make_clust_col}()},
\code{\link{make_rand_cent}()}

Other centroid_functions: 
\code{\link{calc_member_dist_cent}()},
\code{\link{check_clust_cent}()}

Other dbscan_functions: 
\code{\link{cluster_dbscan}()}
}
\concept{centroid_functions}
\concept{clustering_components}
\concept{dbscan_functions}
