% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_dist_col.R
\name{calc_clust_dist_col}
\alias{calc_clust_dist_col}
\title{Calculate the distance between the snps and the cluster centroid}
\usage{
calc_clust_dist_col(c_num, data_mat, centroids_df, norm_typ = "F")
}
\arguments{
\item{c_num}{The clust number to calculate the distance to}

\item{data_mat}{the data matrix used for the coordinate of the snp.
Rows are snps, columns are the trait axes.}

\item{centroids_df}{The dataframe of the centroid co-ordinates for each
cluster. The rows are each cluster number and the columns are the traits.}

\item{norm_typ}{The type of norm to use in the distance calculation. The
default is the Froebenius norm "F".}
}
\value{
clust_dist_df
}
\description{
Calculate the distance between each point in "b_mat"
  and the cluster centroid allocated to c_num.
  This is stored in a dataframe "clust_dist_df" which has rownames for
  the snps "snp_id", and columns "clust_i", for each column.
}
\seealso{
Other distance_functions: 
\code{\link{calc_all_dist_pairs}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_col_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_pair_dist_df}()},
\code{\link{calc_snp_cent_dist}()},
\code{\link{max_dist_calc}()}

Other centroid functions: 
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_snp_cent_dist}()}

Other clustering_components: 
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_snp_cent_dist}()},
\code{\link{check_clust_cent}()},
\code{\link{mat_to_angle_mat}()}
}
\concept{centroid functions}
\concept{clustering_components}
\concept{distance_functions}
