% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_prob.R
\name{calc_clust_prob}
\alias{calc_clust_prob}
\title{Calculate the probability the snp is in the cluster}
\usage{
calc_clust_prob(cluster_df)
}
\arguments{
\item{cluster_df}{The cluster dataframe.}
}
\value{
p
}
\description{
\deqn{p=\frac{1.0}{1.0+d}}
}
\seealso{
Other clustering_components: 
\code{\link{calc_clust_cent}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{check_clust_cent}()},
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{get_aic}()},
\code{\link{km_nan}()},
\code{\link{make_clust_col_name}()},
\code{\link{make_clust_col}()},
\code{\link{make_rand_cent}()}
}
\concept{clustering_components}
\concept{probability_functions}
