% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_col_dist.R
\name{calc_col_dist}
\alias{calc_col_dist}
\title{Find the distance between a snp and all other snps.}
\usage{
calc_col_dist(score_mat, snp1, norm_typ = "F")
}
\arguments{
\item{score_mat}{Matrix of scores for each snp}

\item{snp1}{the snp name to compare all others to}

\item{norm_typ}{the type of norm to use in distance calculations.
default is the Froebenius norm "F".}
}
\value{
dist_df
}
\description{
Use [calc_pair_dist_df] to find the distance between pairs of snps.
  The results from this for each pair is bound on the rows using
  "rbind" into the datframe "dist_df".
}
\seealso{
Other distance_functions: 
\code{\link{calc_all_dist_pairs}()},
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_pair_dist_df}()},
\code{\link{calc_snp_cent_dist}()},
\code{\link{max_dist_calc}()}
}
\concept{distance_functions}
