% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_col_scale.R
\name{calc_col_scale}
\alias{calc_col_scale}
\title{Scale the values in the column of a matrix around the mean}
\usage{
calc_col_scale(col, data, mu, se)
}
\arguments{
\item{col}{Column to be scaled}

\item{data}{Matrix of data}

\item{mu}{Mean of the column}

\item{se}{Standard error of the column}
}
\value{
data_out
}
\description{
Check if SE is 0, if 0 then all points are set to 0.
  If not 0 then rescale using (x-mu)/se
}
\details{
if (se==0){data_out = 0}
  else{ data_out = (data[,col]-mu)/se}
}
\seealso{
Other pca_functions: 
\code{\link{calc_scale_mat}()}
}
\concept{pca_functions}
