% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_member_dist_cent.R
\name{calc_member_dist_cent}
\alias{calc_member_dist_cent}
\title{Calculate the distance between the snps and their allocated cluster centroid}
\usage{
calc_member_dist_cent(snp_id, b_mat, cluster_df, centroids_df, norm_typ = "F")
}
\arguments{
\item{snp_id}{the label for the snp}

\item{b_mat}{the data matrix used for the coordinate of the snp.
Rows are snps, columns are the trait axes.}

\item{cluster_df}{The dataframe with the snp cluster memberships.
Columns are\:
 * "clust_num" the cluster number.
 * "clust_dist" distance from the snp to the cluster centroid.
 * "clust_prob" probability the snp is correctly allocated to the cluster.
Rows are each snp.}

\item{centroids_df}{The dataframe of the centroid co-ordinates for each
cluster. The rows are each cluster number and the columns are the traits.}

\item{norm_typ}{The type of norm to use in the distance calculation. The
default is the Froebenius norm "F".}
}
\value{
snp_clust_df
}
\description{
Calculate the distance between the point in "b_mat"
  corresponding to "snp_id" and the cluster centroid allocated to "snp_id".
  This is stored in a dataframe "snp_clust_df" which has rowname "snp_id",
  and column "clust_dist", with the calculated distance.
}
\seealso{
Other distance_functions: 
\code{\link{calc_all_clust_dist}()},
\code{\link{find_closest_clust_snp}()}

Other centroid_functions: 
\code{\link{calc_clust_cent}()},
\code{\link{check_clust_cent}()}

Other clustering_components: 
\code{\link{calc_clust_cent}()},
\code{\link{calc_clust_prob}()},
\code{\link{check_clust_cent}()},
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{get_aic}()},
\code{\link{km_nan}()},
\code{\link{make_clust_col_name}()},
\code{\link{make_clust_col}()},
\code{\link{make_rand_cent}()}
}
\concept{centroid_functions}
\concept{clustering_components}
\concept{distance_functions}
