% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pair_dist_df.R
\name{calc_pair_dist_df}
\alias{calc_pair_dist_df}
\title{Calculate the distance between a pair of snps and return dataframe}
\usage{
calc_pair_dist_df(score_mat, snp1, snp2, norm_typ = "F")
}
\arguments{
\item{score_mat}{the data matrix. Rows correspond to snps.}

\item{snp1}{the first snp name to use.}

\item{snp2}{the second snp name to use}

\item{norm_typ}{the type of norm to use in the distance calculation.
The default is the Froebenius norm "F".}
}
\value{
dist_df
}
\description{
The distance between the snps is given by the norm of the difference
  between their co-ordinates. With the co-ordinates being the corresponding
  rows in the "score_mat". The type of norm used given by the "norm_typ".
  NaNs are ignored so snps are only compared on common axis.
  The output dataframe "dist_df" is 1 row with
    * "snp1" the name of the first snp
    * "snp2" the name of the second snp
    * "dist" the distance between the snps.
}
\seealso{
Other distance_functions: 
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_snp_cent_dist}()},
\code{\link{max_dist_calc}()}
}
\concept{distance_functions}
