% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snp_cent_dist.R
\name{calc_snp_cent_dist}
\alias{calc_snp_cent_dist}
\title{Calculate the distance between the snps and their allocated cluster centroid}
\usage{
calc_snp_cent_dist(snp_id, data_mat, cent, norm_typ = "F")
}
\arguments{
\item{snp_id}{the label for the snp}

\item{norm_typ}{The type of norm to use in the distance calculation. The
default is the Froebenius norm "F".}

\item{b_mat}{the data matrix used for the coordinate of the snp.
Rows are snps, columns are the trait axes.}

\item{cluster_df}{The dataframe with the snp cluster memberships.
Columns are\:
 * "clust_num" the cluster number.
 * "clust_dist" distance from the snp to the cluster centroid.
 * "clust_prob" probability the snp is correctly allocated to the cluster.
Rows are each snp.}

\item{centroids_df}{The dataframe of the centroid co-ordinates for each
cluster. The rows are each cluster number and the columns are the traits.}
}
\value{
snp_clust_df
}
\description{
Calculate the distance between the point in "b_mat"
  corresponding to "snp_id" and the cluster centroid allocated to "snp_id".
  This is stored in a dataframe "snp_clust_df" which has rowname "snp_id",
  and column "clust_dist", with the calculated distance.
}
\seealso{
Other distance_functions: 
\code{\link{calc_all_dist_pairs}()},
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_col_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{calc_pair_dist_df}()},
\code{\link{max_dist_calc}()}

Other centroid functions: 
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()}

Other clustering_components: 
\code{\link{calc_clust_dist_col}()},
\code{\link{calc_clust_dist}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{check_clust_cent}()},
\code{\link{mat_to_angle_mat}()}
}
\concept{centroid functions}
\concept{clustering_components}
\concept{distance_functions}
