% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_sq_clusts.R
\name{calc_sum_sq_clusts}
\alias{calc_sum_sq_clusts}
\title{Sum the squares of the distances for a group of terms}
\usage{
calc_sum_sq_clusts(num, data, dist_col, group_col)
}
\arguments{
\item{num}{The number to group the terms on}

\item{data}{The data to take the values from}

\item{dist_col}{The name of the column to get the values from}

\item{group_col}{The name of the column to group terms on}
}
\value{
sum of the squares
}
\description{
for all the terms in group_col with value num sum the
  square of the dist_col values
}
\seealso{
Other cluster_properties: 
\code{\link{calc_clust_ic}()},
\code{\link{calc_clust_prob}()},
\code{\link{calc_tot_within}()}

Other ic_functions: 
\code{\link{calc_tot_within}()},
\code{\link{find_all_ic}()},
\code{\link{test_find_aic}()}
}
\concept{cluster_properties}
\concept{ic_functions}
