% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clust_cent.R
\name{check_clust_cent}
\alias{check_clust_cent}
\title{Check convergence for the cluster centroid}
\usage{
check_clust_cent(
  c_num,
  clustnum_df,
  b_mat,
  centroids_df,
  norm_typ = "F",
  clust_threshold = 1e-05
)
}
\arguments{
\item{c_num}{The number of the cluster to be checked}

\item{clustnum_df}{The cluster membership dataframe. The rows are the snps,
The columns are "clust_num".}

\item{b_mat}{The data for each snp. Each row is the snps coordinate.}

\item{centroids_df}{The dataframe of the centroids for each cluster on
the previous iteration. The rows are the cluster numbers and the
columns are the centroid value on that trait axis.}

\item{norm_typ}{String indicating the type of norm to be used when
calculating the distance between the centroids. Default is the Froebenius
norm "F".}

\item{clust_threshold}{The threshold for how close the clusters should
be to be considered converged. default \:1e-5}

\item{na_rm}{Boolean. If TRUE (default) then remove NaNs in means.}
}
\value{
out_centroid_df
}
\description{
Check convergence for the cluster centroid
}
\details{
The check can only be made if there are snps which are members of the
  cluster.
  If the cluster is not empty then the new centres are found using
    the means of the snps in the cluster.
  If the new cluster centroid is within "clust_threshold" of the previous
    centroid then "thresh_check"=TRUE.
  If thresh_check then create a dataframe "out_centroid_df" with the
    previous centroid and the column "thesh_check" with TRUE in it.
  else create a dataframe "out_centroid_df" of the new centroids and
    the column "thresh_check" with FALSE in it.
}
\seealso{
Other clustering_components: 
\code{\link{calc_member_dist_cent}()},
\code{\link{convert_mat_to_angle_mat}()}
}
\concept{centroid_functions}
\concept{clustering_components}
