% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_pca_compare_iterative.R
\name{clust_pca_compare_iterative}
\alias{clust_pca_compare_iterative}
\title{Cluster data on the principal components and score the clusters on
association. Compare the cluster scores and iteratively add a
trait each time.}
\usage{
clust_pca_compare_iterative(
  data_matrices,
  out_pheno,
  thresholds = list(threshmul = 5, diff = 1e-05, clust = 1e-05),
  na_handling = list(narm = TRUE, percent = 0.95),
  iter_traits = list(iter = 0, bp_on = TRUE, clust_prob_on = TRUE, clust_typ = "basic"),
  norm_typs = list(clust = "F", thresh = "F"),
  nums = list(max_dist = 1, np = 3, nr = 5)
)
}
\arguments{
\item{data_matrices}{- A list containing the matrices
* "beta" matrix corresponding to the data.
* "pval" the matrix of probabilities associated with each score.
* "se" matrix of standard errors associated with each score.
* "trait_info" which is the dataframe of all the traits.}

\item{out_pheno}{- Label for the outcome phenotype}

\item{thresholds}{- List for the three threshold related variables.
* "threshmul" is multiplied by the data variance for cluster
difference threshold.
* "diff" is the the cluster difference threshold once found.
* "clust" is the required distance between cluster centres for a
cluster to be considered converged.}

\item{na_handling}{- List of the terms related to handling NaNs
* "narm" - TRUE (default) if NaNs are to be removed in calculations.
* "percent" - percentage of a column which must be not NaN. default\:0.95}

\item{iter_traits}{- list of terms indicating the type of program.
* "iter" - integer, (default 0)
* "bp_on" - TRUE (default) if probability of scores is to be used.
* "clust_prob_on" - TRUE (default) if prob of being in cluster to be used.
* "clust_typ" - default="basic", the clustering method to use.}

\item{norm_typs}{- List of norm types
* "clust" - default="F"
* "thresh" - default="F"}

\item{nums}{- List of important numbers.
* "max_dist" - Maximum distance between points, default=1
* "np" - Number of principal components
* "nc" - Number of clusters.}
}
\value{
df_list
}
\description{
df_list is the list of of the clusters, the principal component matrices,
  cluster scores, and the maximum difference between clusters.
  Each iteration generated using [clust_pca_compare_single].
}
\seealso{
Other cluster_wrappers: 
\code{\link{clust_pca_all}()},
\code{\link{clust_pca_compare_single}()},
\code{\link{full_cluster_and_plot}()}
}
\concept{cluster_wrappers}
