% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_pca_compare_single.R
\name{clust_pca_compare_single}
\alias{clust_pca_compare_single}
\title{Single iteration for iterating through trait list.}
\usage{
clust_pca_compare_single(
  df_list,
  iter_traits,
  num_axis,
  data_matrices,
  na_handling,
  thresholds,
  norm_typs,
  nums
)
}
\arguments{
\item{df_list}{the results from previous iterations to append to.}

\item{iter_traits}{- list of terms indicating the type of program.
* "iter" - integer, (default 0)
* "bp_on" - TRUE (default) if probability of scores is to be used.
* "clust_prob_on" - TRUE\:default switch for using prob of being in cluster
* "clust_typ" - default\:"basic", the clustering method to use.}

\item{num_axis}{Number of trait axes in use.}

\item{data_matrices}{- A list containing
* "beta" main data matrix
* "pval" probabilities associated with "beta" values
* "se"  standard errors associated with "beta" values.
* "trait_info" data frame of all the traits.}

\item{na_handling}{- List of the terms related to handling NaNs
* "narm" - TRUE (default) if NaNs are to be removed in calculations.
* "percent" - percentage of a column which must be not NaN. default\:0.95}

\item{thresholds}{- List of threshold related variables.
* "threshmul" is multiplied by the data variance for cluster
difference threshold.
* "diff" is the the cluster difference threshold once found.
* "clust" is the required distance between cluster centres for a
cluster to be considered converged.}

\item{norm_typs}{- List of norm types
* "clust" - default\:"F"
* "thresh" - default\:"F"}

\item{nums}{- List of important numbers.
* "max_dist" - Maximum distance between points, default=1
* "np" - Number of principal components
* "nc" - Number of clusters.}
}
\value{
df_list
}
\description{
Single iteration for iterating through trait list.
}
\details{
df_list is the list of of the clusters, the principal component matrices,
  cluster scores, and the maximum difference between clusters.
  Each iteration generated using [clust_pca_compare_single].
}
\seealso{
Other cluster_wrappers: 
\code{\link{clust_pca_all}()},
\code{\link{clust_pca_compare_iterative}()},
\code{\link{full_cluster_and_plot}()}
}
\concept{cluster_wrappers}
