% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_dbscan.R
\name{cluster_dbscan}
\alias{cluster_dbscan}
\title{Cluster the data using kmeans then minimising aic.}
\usage{
cluster_dbscan(data_matrices, eps = 0.4, bin_p_clust = FALSE)
}
\arguments{
\item{data_matrices}{List of the matrices of data.
Conatins the\:
  * "beta" initial matrix of assocaition data
  * "se" initial matrix of standard error data
  * "pval" initial matrix of p-value data
  * "trait_info" list of informations about the traits
  * "beta_pc" reformatted matrix of assocaition data
  * "se" reformatted matrix of standard error data
  * "pval" reformatted matrix of p-value data
  * "transform" matrix used for the pca transformation
  * Rows correspond to snps, columns to traits}

\item{eps}{The density deining a cluster. default \: 0.4}

\item{bin_p_clust}{Should the cluster centroid be weighted by p-values. default \: FALSE}
}
\value{
clusters_df
}
\description{
If space_typ == "angle" then data is converted to angles.
  for i=1:(nr+1)
     The points are clustered using [km_nan] and i clusters
     find the aic for each cluster using [find_ic].
  The set of clusters which minimises the aic is chosen.
  Outputs: clustout = [clusters, clust_dist, centres]
  The "clusters_df" dataframe of the cluster membership with columns\:
    * "clust_num" the number of clusters
    * "clust_dist" the distance from the snp to the cluster centre
      (or distance between angles)
    * "clust_prob" probability the snp is in the cluster. Calculated
      using [calc_clust_prob].
  The clust_dist dataframe has columns corresponding to the clusters,
    the rows are the datapoints and each value is the distance from that
    point to the cluster centre.
  The centres dataframe has columns corresponding to the axis and the rows
    corresponding to the clusters. Each row is the co-ordinate of the cluster
    centroid.
}
\seealso{
Other cluster_functions: 
\code{\link{cluster_kmeans_min}()},
\code{\link{cluster_kmeans}()},
\code{\link{get_aic}()},
\code{\link{km_nan}()}

Other dbscan_functions: 
\code{\link{calc_clust_cent}()}
}
\concept{cluster_functions}
\concept{dbscan_functions}
