% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_program.R
\name{clustering_program}
\alias{clustering_program}
\title{Program for preparing data then clustering it.}
\usage{
clustering_program(iter_traits, test = 0)
}
\arguments{
\item{iter_traits}{dataframe of the indicators for the types of
calculations that will be done.
Contains\:
* `dname`\: data directory name
* `clust_type`\: name of the clustering method
* `pc_type`\: method for finding principal components,
             options are "manual", "prcomp", "none"
* `n_clust`\: number of clusters
* `n_pcs`\: number of PCs
* `bin_angles`\: 1 or 0 use angles
* `bin_p_clust`\: 1 or 0 use p-vals in clustering
* `bin_p_score`\: 1 or 0 use p-vals in score
* `bin_d_score`\: 1 or 0 use cluster distance for cluster scoring
* `point_eps` \: the density of points for density based methods. default \:0.4
* `nan_rm`\: how to handle NaN values. If 1 remove and use pairwise.}

\item{test}{Indicator for whether to run as a test or full program.
default \: 1 (test case)}
}
\description{
Program for preparing data then clustering it.
}
