% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_mat_colnames.R
\name{crop_mat_colnames}
\alias{crop_mat_colnames}
\title{Crop matrix using column names}
\usage{
crop_mat_colnames(mat, num_rows, col_names)
}
\arguments{
\item{mat}{matrix to be cropped}

\item{num_rows}{the number of rows to keep}

\item{col_names}{the column names to include.}
}
\description{
Return the matrix with row 1:num_rows
  and column names in col_names
}
