% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_matrices.R
\name{crop_mat_list}
\alias{crop_mat_list}
\title{Crop all of the matrices in mat_list}
\usage{
crop_mat_list(mat_list, trait_df, out_pheno, exp_pheno, n_rows, n_col0, n_col1)
}
\arguments{
\item{mat_list}{The list of matrices that need to be cropped.}

\item{trait_df}{The dataframe of the traits.}

\item{out_pheno}{The outcome phenotype label}

\item{exp_pheno}{The ecposure phenotype label.}

\item{n_rows}{The number of rows to crop by.}

\item{n_col0}{The first column to include}

\item{n_col1}{The last column to include.}
}
\value{
List of cropped matrices.
}
\description{
include only the first n_rows
  and columns between n_col0 and n_col1,
  ensure that out_pheno and exp_pheno are included.
}
