% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_matrices.R
\name{crop_matrices}
\alias{crop_matrices}
\title{Crop all of the matrices in mat_list}
\usage{
crop_matrices(mat_list, trait_df, out_pheno, exp_pheno, n_rows, n_col0, n_col1)
}
\arguments{
\item{mat_list}{The list of matrices that need to be cropped.}

\item{trait_df}{The dataframe of the traits.}

\item{out_pheno}{The outcome phenotype label}

\item{exp_pheno}{The ecposure phenotype label.}

\item{n_rows}{The number of rows to crop by.}

\item{n_col0}{The first column to include}

\item{n_col1}{The last column to include.}
}
\value{
List of cropped matrices.
}
\description{
include only the first n_rows
  and columns between n_col0 and n_col1,
  ensure that out_pheno and exp_pheno are included.
}
\details{
1. Get the columns of each matrix between n_col0 and n_col1.
    Crop to n_rows at this step at the same time
  2. Check if the exposure and outcome columns are in the n_col range.
  3. If not then:
     a. Get the matrix with just the exposure and outcome columns
     using `crop_mat_colnames`. Crop to n_rows at this step at the same time
  4. Merge the two sets to create a matrix with columns in the n_col
  range and the exposure and outcome columns.
}
\seealso{
Other preconditioning_functions: 
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{convert_point_to_angle}()},
\code{\link{crop_mat_colnames}()},
\code{\link{crop_mat_colnums}()},
\code{\link{reformat_data}()},
\code{\link{rescale_by_end_col}()}
}
\concept{preconditioning_functions}
