% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_clust_snp.R
\name{find_closest_clust_snp}
\alias{find_closest_clust_snp}
\title{Find the cluster whose centroid is closest to snp_id}
\usage{
find_closest_clust_snp(snp_id, b_mat, cluster_df, centroids_df, max_dist = 10)
}
\arguments{
\item{snp_id}{the label for the snp}

\item{b_mat}{the matrix of the data. The rows correpond to the snps.}

\item{cluster_df}{dataframe of cluster membership. Each row is a snp,
it has columns for\:
 * "clust_num" the number of the allocated cluster.
 * "clust_dist" the distane to the centre of the cluster.
 * "clust_prob" probability the snp is in the assigned cluster.}

\item{centroids_df}{dataframe of the centroid co-ordinates. Each row
corresponds to a cluster number and the columns are the trait axes.}

\item{max_dist}{The maximum distance between any two points. Used
for initialisation if snp is not already in the cluster assignment.}
}
\value{
out_list
}
\description{
Find the distance between the snp and each cluster centroid.
  If the distance is less then the currently assigned cluster then
  reassign the cluster number and cluster distance.
}
\details{
The dataframe "snp_clust_df"  has a row labelled by the "snp_id".
  The columns are\:
  * "clust_num" The number of the closest cluster
  * "clust_dist" The distance from the snp to the centroid of that cluster.
  * "clust_prob" probability the snp is in the cluster it's been assigned.
  The dataframe "snp_clust_dist_df"  has a row labelled by the "snp_id".
  The columns are the clust numbers, each value is the distance from the
  snp to that cluster.

  out_list = ("clusters" = snp_clust_df, "clust_dist" = snp_clust_dist_df)
}
\seealso{
Other distance_functions: 
\code{\link{calc_member_dist_cent}()}
}
\concept{distance_functions}
