% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_low_percent_col.R
\name{find_low_percent_col}
\alias{find_low_percent_col}
\title{Find rownames corresponding to the lowest `col_percent` of the column `col`}
\usage{
find_low_percent_col(data_mat, col, col_percent = 0.1, na_rm = TRUE)
}
\arguments{
\item{data_mat}{the data matrix}

\item{col}{the column label}

\item{col_percent}{the threshold percentage}

\item{na_rm}{whether to remove NaNs or not}
}
\description{
Find rownames corresponding to the lowest `col_percent` of the column `col`
}
