% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_low_percent_col_mat.R
\name{find_low_percent_col_mat}
\alias{find_low_percent_col_mat}
\title{Find the rownames corresponding to the lowest `col_percent` of each column}
\usage{
find_low_percent_col_mat(data_mat, col_percent, na_rm = TRUE)
}
\arguments{
\item{data_mat}{the data matrix}

\item{col_percent}{the threshold percentage}

\item{na_rm}{whether to remove NaNs or not}
}
\description{
Find the rownames corresponding to the lowest `col_percent` of each column
}
