% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aic.R
\name{get_aic}
\alias{get_aic}
\title{Get the set of clusters which minimizes aic}
\usage{
get_aic(clust_out)
}
\arguments{
\item{clust_out}{list of cluster dataframes for number of clusters}
}
\description{
Get the set of clusters which minimizes aic
}
\seealso{
Other clustering_components: 
\code{\link{calc_clust_cent}()},
\code{\link{calc_clust_prob}()},
\code{\link{calc_member_dist_cent}()},
\code{\link{check_clust_cent}()},
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{km_nan}()},
\code{\link{make_clust_col_name}()},
\code{\link{make_clust_col}()},
\code{\link{make_rand_cent}()}

Other cluster_functions: 
\code{\link{cluster_dbscan}()},
\code{\link{cluster_kmeans_min}()},
\code{\link{cluster_kmeans}()},
\code{\link{km_nan}()}
}
\concept{cluster_functions}
\concept{clustering_components}
