% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_mat_on_name.R
\name{join_mat_on_name}
\alias{join_mat_on_name}
\title{Take the matrices with matching label name from each list
and column bind them}
\usage{
join_mat_on_name(name, mat_list_a, mat_list_b)
}
\arguments{
\item{name}{The name of the matrix in each list.}

\item{mat_list_a}{first matrix list, includes a matrix labelled name}

\item{mat_list_b}{second matrix list, includes a matrix labelled name}
}
\value{
matrix combine the matrix named name in each list. Bound on columns.
}
\description{
Take the matrices with matching label name from each list
and column bind them
}
