% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_nan.R
\name{km_nan}
\alias{km_nan}
\title{Apply kmeans clustering ignoring NaN terms}
\usage{
km_nan(
  b_mat,
  nclust = 5,
  iter_max = 300,
  clust_threshold = 1e-05,
  norm_typ = "F",
  na_rm = TRUE,
  prob_on = TRUE,
  how_cents = "point"
)
}
\arguments{
\item{b_mat}{Matrix of the data, each row is a snp and the columns
are the traits.}

\item{nclust}{The number of clusters to allocate. default\:5}

\item{iter_max}{The maximum number of iterations to reach cluster
convergence. default\:300}

\item{clust_threshold}{The threshold for how close cluster
centres need to be to be considered converged. default\:1e-5}

\item{norm_typ}{The type of norm to be used in distance calculation.
The default is the Froebenius norm "F".}

\item{na_rm}{Bool to indicate how to deal with NaNs. If TRUE (default)
then NaNs are ignored.}

\item{prob_on}{Bool to indicate whether cluster probability is to be
assign. default\:TRUE}

\item{how_cents}{How the centroids will be initialised. If "rand (default)"
then the coordinates are created using a uniform distribution on the range
of each axis. If "points" then the centroid coordinates are assign to
nclust random points from the dataspace.}
}
\value{
clust_out
}
\description{
Use common axes for distance calculations to igore NaNs and
retain data.
}
\details{
The cluster centres are randomly assigned using [make_rand_cent].
The cluster centres are check and reassigned using [check_clust_cent],
When the clusters are converged or the maximum number of iterations are
used then the clusters are return in the dataframe "cluster_df".
"clust_out" is the list containing the "clusters_df" dataframe
labelled "cluster" of the cluster membership with columns\:
  * "clust_num" the number of clusters
  * "clust_dist" the distance from the snp to the cluster centre
  (or distance between angles)
  * "clust_prob" probability the snp is in the cluster. Calculated
  using [calc_clust_prob].
"clust_out" also contains the "centroids_df" dataframe labelled "centres"
whose columns are the traits and rows are the cluster numbers. Additional
column "thresh_check" contains Bool indicating whether that cluster
converged.
}
