% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clust_col_name.R
\name{make_clust_col_name}
\alias{make_clust_col_name}
\title{make_clust_col_name}
\usage{
make_clust_col_name(c_num)
}
\arguments{
\item{c_num}{integer for cluster number}
}
\value{
"clust_"c_num
}
\description{
Function for creating the cluster label.
}
\seealso{
Other label_functions: 
\code{\link{make_pc_col_name}()}
}
\concept{label_functions}
