% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_path_label_str.R
\name{make_path_label_str}
\alias{make_path_label_str}
\title{Create the string representing the variations in the iterations.}
\usage{
make_path_label_str(iter_traits)
}
\arguments{
\item{iter_traits}{- Data frame containing the iteration variables
"bp_on", "clust_prob_on",  "clust_typ"}
}
\value{
label_str
}
\description{
This string will have no space so it is suitable for file and
 directory names.
}
\details{
The string will start with the clustering type, then the bp switch,
then the cluster probability switch.
}
\seealso{
Other setup_functions: 
\code{\link{setup_matrices}()}
}
\concept{setup_functions}
