% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pc_col.R
\name{make_pc_col}
\alias{make_pc_col}
\title{Create a dataframe with column Pi.}
\usage{
make_pc_col(i, rows = character())
}
\arguments{
\item{i}{- The number for the column label}

\item{rows}{- The rows for the dataframe,
defalut\: character() for an empty dataframe}
}
\value{
dataframe(Pi = numeric())
}
\description{
Create a dataframe with column Pi.
}
\details{
i is the input number
}
