% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pc_col_name.R
\name{make_pc_col_name}
\alias{make_pc_col_name}
\title{make_pc_col_name}
\usage{
make_pc_col_name(i)
}
\arguments{
\item{i}{integer for PC}
}
\value{
"PC"c_num
}
\description{
Function for creating the cluster label.
}
\seealso{
Other label_functions: 
\code{\link{make_clust_col_name}()}
}
\concept{label_functions}
