% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clust_angle_scatter.R
\name{plot_clust_angle_scatter}
\alias{plot_clust_angle_scatter}
\title{Plot the scatter plot or snps and principal components.}
\usage{
plot_clust_angle_scatter(
  cluster_df,
  b_mat,
  se_mat,
  iter_traits,
  col1,
  col2,
  num_axis = 2,
  pw = 8,
  ph = 4
)
}
\arguments{
\item{cluster_df}{The dataframe of cluster_df membership}

\item{b_mat}{The matrix of the score data}

\item{se_mat}{The matrix of the standarad errors associated with the scores.}

\item{iter_traits}{The iteration variables for the type of iteration.}

\item{num_axis}{The number of trait axis, default\:0}

\item{pw}{The plot width, default\:8}

\item{ph}{The plot heigh, default\:4}

\item{exp_pheno}{The label for the exposure phenotype}

\item{out_pheno}{The label for the output phenotype}
}
\description{
Colour by cluster number.
The x and y axis are the assocation with the principal component
The width and height of the errorbars are given by the standard error.
}
