% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clust_scatter_test.R
\name{plot_clust_scatter_test}
\alias{plot_clust_scatter_test}
\title{Plot the scatter plot or snps and principal components.}
\usage{
plot_clust_scatter_test(
  cluster_df,
  b_mat,
  iter_traits,
  c1,
  c2,
  num_axis = 2,
  pw = 8,
  ph = 4,
  save_suffix = ""
)
}
\arguments{
\item{cluster_df}{The dataframe of cluster_df membership}

\item{b_mat}{The matrix of the score data}

\item{iter_traits}{The iteration variables for the type of iteration.}

\item{c1}{The column for the x-axis}

\item{c2}{The column for the y-axis}

\item{num_axis}{The number of trait axis, default\:0}

\item{pw}{The plot width, default\:8}

\item{ph}{The plot heigh, default\:4}

\item{save_suffix}{String to add to plot save name. default \: ""}
}
\description{
Colour by cluster number.
The x and y axis are the assocation with the principal component
The width and height of the errorbars are given by the standard error.
}
