% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot the scatter plot or snps and principal components.}
\usage{
plot_scatter(b_mat, iter_traits, num_axis = 2, pw = 8, ph = 4)
}
\arguments{
\item{b_mat}{The matrix of the score data}

\item{iter_traits}{The iteration variables for the type of iteration.}

\item{num_axis}{The number of trait axis, default\:0}

\item{pw}{The plot width, default\:8}

\item{ph}{The plot heigh, default\:4}
}
\description{
Colour by cluster number.
The x and y axis are the assocation with the principal component
The width and height of the errorbars are given by the standard error.
}
