% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_data}
\alias{reformat_data}
\title{Reformat the data into the structure desired for clustering.}
\usage{
reformat_data(data_matrices, bin_angles = 1, pca_type = "none", np = 1)
}
\arguments{
\item{data_matrices}{The list of matrices to be transformed.
The data governing any transformation is in the "beta" matrix}

\item{bin_angles}{binary variable indicating whether to convert to
angle data or not. 1 is angles, 0 is not.}

\item{pca_type}{Method for principal components to use. Options are
"manual", "prcomp" or "none"}

\item{np}{The number of principal components to find}
}
\value{
data_matrices List of matrices
}
\description{
1. If the angles switch is set convert the data to the angles.
    The is only done for the main beta data matrix.
  2. Find the principal components for the beta matrix. Use the method set
    in pca_type, if none then skip and return the data_matrices.
}
\seealso{
Other preconditioning_functions: 
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{convert_point_to_angle}()},
\code{\link{crop_mat_colnames}()},
\code{\link{crop_mat_colnums}()},
\code{\link{crop_matrices}()},
\code{\link{rescale_by_end_col}()}
}
\concept{preconditioning_functions}
