% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_by_end_col.R
\name{rescale_by_end_col}
\alias{rescale_by_end_col}
\title{Multiply all by the final terms in a row by the final term.
Repeat for all rows.}
\usage{
rescale_by_end_col(mat)
}
\description{
Multiply all by the final terms in a row by the final term.
Repeat for all rows.
}
\concept{utility}
