% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_matrices.R
\name{setup_matrices}
\alias{setup_matrices}
\title{read in data from csv files as matrices.}
\usage{
setup_matrices(
  data_dir,
  test = 0,
  num_rows = 0,
  num_trait0 = 0,
  num_trait1 = 0
)
}
\arguments{
\item{data_dir}{the directory location of the data}

\item{test}{Switch 0,1. If 1 then crop the data with the
test variables. default\:0}

\item{num_rows}{If test=1 then this is the number of rows
of the data that will be used. default\:0}

\item{num_trait0}{if test==1 then this is the first column
that will be used. default\:0}

\item{num_trait1}{if test==1 then this is the last column
that will be used. default\:0}
}
\description{
The directory must contain files with the names\:
  * "pval_df.csv"
  * "unstdBeta_df.csv"
  * "unstdSE_df.csv"
  * "trait_info_nfil.csv"
 If test parameters are set the data will be cropped to run
the functions on a smaller dataset.
}
\details{
rows labeled by SNP_id and columns labelled by trait info.
The entries at each position correspond to the values for beta, se, t, and p
}
