% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add_col_df_list.R
\name{test_add_col_df_list}
\alias{test_add_col_df_list}
\title{Test add a column with value to a list of dataframe}
\usage{
test_add_col_df_list()
}
\description{
Test add a column with value to a list of dataframe
}
