% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add_nclust_cols_empty.R
\name{test_add_nclust_cols_empty}
\alias{test_add_nclust_cols_empty}
\title{Add "np" columns to a dataframe.}
\usage{
test_add_nclust_cols_empty()
}
\description{
Create an empty dataframe.
Check that the column names after add_nclust_cols(5)
are (C1,C2,C3,C4,C5)
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_calc_clust_dist_col}()},
\code{\link{test_calc_clust_dist}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_find_aic}()},
\code{\link{test_find_low_percent_col_mat}()},
\code{\link{test_find_low_percent_col}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}
}
\concept{tests}
