% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add_nclust_cols_onecol.R
\name{test_add_nclust_cols_onecol}
\alias{test_add_nclust_cols_onecol}
\title{Add "np" columns to a dataframe.}
\usage{
test_add_nclust_cols_onecol()
}
\description{
Create an empty dataframe with column A.
Check that the column names after add_nclust_cols(5)
are (A, C1,C2,C3,C4,C5)
}
\seealso{
Other test: 
\code{\link{test_add_np_cols_onecol}()}
}
\concept{test}
