% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add_np_cols_onecolrow.R
\name{test_add_np_cols_onecolrow}
\alias{test_add_np_cols_onecolrow}
\title{Add "np" columns to a dataframe.}
\usage{
test_add_np_cols_onecolrow()
}
\description{
Create an empty dataframe with column A.
Check that the column names after add_nclust_cols(5)
are (A, P1,P2,P3,P4,P5)
}
\seealso{
Other dataframe_editors: 
\code{\link{add_nclust_cols}()},
\code{\link{add_np_cols}()},
\code{\link{test_add_nclust_cols_onecolrow}()}
}
\concept{dataframe_editors}
