% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add_np_cols_onerow.R
\name{test_add_np_cols_onerow}
\alias{test_add_np_cols_onerow}
\title{Add "np" columns to a dataframe.}
\usage{
test_add_np_cols_onerow()
}
\description{
Create an empty dataframe.
Check that the column names after add_npcols(5)
are (P1,P2,P3,P4,P5)
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_find_aic}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}
}
\concept{tests}
