% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na_col_check.R
\name{test_all_na}
\alias{test_all_na}
\title{Test whether the function for checking the NaNs in a column works.}
\usage{
test_all_na(b_mat, nan_col = "30600_irnt")
}
\arguments{
\item{b_mat}{data matrix}

\item{nan_col}{The column name to check for NaNs. default\:"30600_irnt"
Should return 0}
}
\value{
1 if succesful, 0 if not
}
\description{
Test whether the function for checking the NaNs in a column works.
}
