% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_calc_all_clust_dist.R
\name{test_calc_all_clust_dist}
\alias{test_calc_all_clust_dist}
\title{Find the cluster whose centroid is closest to snp_id}
\usage{
test_calc_all_clust_dist()
}
\value{
out_list
}
\description{
Test the function which finds the distance between the snp and
  each cluster centroid. `find_closest_clust_snp`
}
\details{
The dataframe "snp_clust_df"  has a row labelled by the "snp_id".
  The columns are\:
  * "clust_num" The number of the closest cluster
  * "clust_dist" The distance from the snp to the centroid of that cluster.
  * "clust_prob" probability the snp is in the cluster it's been assigned.
  The dataframe "snp_clust_dist_df"  has a row labelled by the "snp_id".
  The columns are the clust numbers, each value is the distance from the
  snp to that cluster.

  out_list = ("clusters" = snp_clust_df, "clust_dist" = snp_clust_dist_df)
}
