% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_calc_clust_cent.R
\name{test_calc_clust_cent}
\alias{test_calc_clust_cent}
\title{Test the function which finds the cluster centroids.}
\usage{
test_calc_clust_cent()
}
\description{
Tests:
* Each cluster number should have a corresponding centroid
* The dimensions of the centroid should be the same as the data matrix.
}
\seealso{
Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_dbscan}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_cluster_kmeans}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_get_aic}()},
\code{\link{test_join_mat_on_name}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
