% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_calc_clust_dist_col.R
\name{test_calc_clust_dist_col}
\alias{test_calc_clust_dist_col}
\title{Test `calc_clust_dist_col` the function which finds the distance
from a point to the centre of a cluster numbered c_num}
\usage{
test_calc_clust_dist_col()
}
\description{
* Check output is dataframe
 * Check the number of rows is same data_mat
 * Check there is one column
 * Check the column name matches the cluster number label
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_calc_clust_dist}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_find_aic}()},
\code{\link{test_find_low_percent_col_mat}()},
\code{\link{test_find_low_percent_col}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}
}
\concept{tests}
