% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_convert_point_to_angle.R
\name{test_convert_point_to_angle}
\alias{test_convert_point_to_angle}
\title{Test the `crop_matrices` function}
\usage{
test_convert_point_to_angle()
}
\description{
Test the `crop_matrices` function
}
\seealso{
Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_cluster_kmeans}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_get_aic}()},
\code{\link{test_join_matlist_on_name}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
