% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_crop_matrices.R
\name{test_crop_matrices}
\alias{test_crop_matrices}
\title{Test the `crop_matrices` function}
\usage{
test_crop_matrices()
}
\description{
Test the `crop_matrices` function
}
\seealso{
Other tests: 
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_join_matlist_on_name}()}
}
\concept{tests}
