% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_find_aic.R
\name{test_find_aic}
\alias{test_find_aic}
\title{Test the function`find_all_ic`}
\usage{
test_find_aic()
}
\arguments{
\item{clust_re}{cluster results dataframe found using [km_nan]}

\item{num_axis}{the number of axis}
}
\value{
clust_re\$clusters
}
\description{
Test the function`find_all_ic`
}
\details{
The aic and bic are found for each cluster using [calc_clust_ic]
  this is added to clust_re\$clusters and clust_re\$clusters is output.
  Ncents is also added as an axis to clust_re\$clusters.
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}

Other ic_functions: 
\code{\link{calc_sum_sq_clusts}()},
\code{\link{calc_tot_within}()},
\code{\link{find_all_ic}()}
}
\concept{ic_functions}
\concept{tests}
