% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_find_low_percent_col_mat.R
\name{test_find_low_percent_col_mat}
\alias{test_find_low_percent_col_mat}
\title{Test for `find_low_percent_col_mat`}
\usage{
test_find_low_percent_col_mat()
}
\description{
Test the function which is supposed to find the
row names for the points in the lowest `col_percent` of each column.

* Check the output is list
* Check the number of terms is 10 percent of the number of rows times
  the number of columns.
* Check the min value is in the list
* check all outputs are in the original row names
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_calc_clust_dist_col}()},
\code{\link{test_calc_clust_dist}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_find_aic}()},
\code{\link{test_find_low_percent_col}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}
}
\concept{tests}
