% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_get_aic.R
\name{test_get_aic}
\alias{test_get_aic}
\title{Test the function for finding the aic of a set of clusters.}
\usage{
test_get_aic()
}
\description{
Test the function for finding the aic of a set of clusters.
}
\details{
Checks:
* there result is a cluster dataframe with an aic column
* the aic is numeric
* the aic is the same on all columns
}
\seealso{
Other tests: 
\code{\link{test_calc_clust_cent}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_dbscan}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_cluster_kmeans}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_join_mat_on_name}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
