% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_join_matlist_on_name.R
\name{test_join_matlist_on_name}
\alias{test_join_matlist_on_name}
\title{Test the `crop_matrices` function}
\usage{
test_join_matlist_on_name()
}
\description{
Test the `crop_matrices` function
}
\seealso{
Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_kmeans}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
