% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na_check.R
\name{test_na_check}
\alias{test_na_check}
\title{Test the `check_col_na` function}
\usage{
test_na_check()
}
\description{
Test the `check_col_na` function
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_calc_clust_dist_col}()},
\code{\link{test_calc_clust_dist}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_find_aic}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()}
}
\concept{tests}
