% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rescale_by_end_col.R
\name{test_rescale_by_end_col}
\alias{test_rescale_by_end_col}
\title{Test the `rescale_by_end_col` function}
\usage{
test_rescale_by_end_col()
}
\description{
Test the `rescale_by_end_col` function
}
\seealso{
Other tests: 
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_join_matlist_on_name}()},
\code{\link{test_reformat_data}()}
}
\concept{tests}
