% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_transform_coords_in_mat.R
\name{test_transform_coords_in_mat}
\alias{test_transform_coords_in_mat}
\title{Test the function `transform_coords_in_mat`}
\usage{
test_transform_coords_in_mat()
}
\value{
vec_mat
}
\description{
* Check the types
* Check the row labels
* Check the column labels
* Check the dimensions
}
