% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clust_prob.R
\name{calc_clust_prob}
\alias{calc_clust_prob}
\title{Calculate the probability the snp is in the cluster}
\usage{
calc_clust_prob(cluster_df)
}
\arguments{
\item{cluster_df}{The cluster dataframe.}
}
\value{
p
}
\description{
\deqn{p=\frac{1.0}{1.0+d}}
}
\concept{clustering functions}
\concept{probability_functions}
