% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmeans_min.R
\name{cluster_kmeans_min}
\alias{cluster_kmeans_min}
\title{Cluster the data using kmeans then minimising aic.}
\usage{
cluster_kmeans_min(
  data_list,
  nclust = 10,
  max_dist = 10,
  space_typ = "regular",
  clust_prob_on = TRUE,
  norm_typ = "F",
  threshold = 1e-05,
  narm = TRUE
)
}
\arguments{
\item{data_list}{List of the matrices of data.
Conatins the\:
  * "beta" matrix of assocaition data
  * "se" matrix of standard error data
  * "pval" matrix of p-value data
  * Rows correspond to snps, columns to traits}

\item{nclust}{The maximum number of clusters to consider. default\:10}

\item{max_dist}{The maximum distance between any two points.}

\item{space_typ}{The spatial structure of the data when clustering.
If "angle" then the data is converted to angles. If "regular" the
data is left the same.}

\item{clust_prob_on}{Bool switch. If TRUE (default) then cluster
probability if used to weight the cluster scores.}

\item{norm_typ}{The type of norm to be used for distance calculations.
The default is the Froebenius norm "F".}

\item{threshold}{The threshold for distance between cluster centres
for clusters to be considered converged.}

\item{narm}{Bool to indicate how to handle NaNs. If TRUE (default)
NaNs are ignored in calculations.}
}
\value{
clusters_df
}
\description{
If space_typ == "angle" then data is converted to angles.
  for i=1:(nr+1)
     The points are clustered using [km_nan] and i clusters
     find the aic for each cluster using [find_ic].
  The set of clusters which minimises the aic is chosen.
  The "clusters_df" dataframe of the cluster membership with columns\:
    * "clust_num" the number of clusters
    * "clust_dist" the distance from the snp to the cluster centre
      (or distance between angles)
    * "clust_prob" probability the snp is in the cluster. Calculated
      using [calc_clust_prob].
}
\concept{cluster_functions}
