% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_mat_colnames.R
\name{crop_mat_colnames}
\alias{crop_mat_colnames}
\title{Crop matrix using column names}
\usage{
crop_mat_colnames(mat, num_rows, col_names)
}
\arguments{
\item{mat}{matrix to be cropped}

\item{num_rows}{the number of rows to keep}

\item{col_names}{the column names to include.}
}
\description{
Return the matrix with row 1:num_rows
  and column names in col_names
}
\seealso{
Other preconditioning_functions: 
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{convert_point_to_angle}()},
\code{\link{crop_mat_colnums}()},
\code{\link{crop_matrices}()},
\code{\link{reformat_data}()},
\code{\link{rescale_by_end_col}()}
}
\concept{preconditioning_functions}
