% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aic.R
\name{get_aic}
\alias{get_aic}
\title{Get the set of clusters which minimizes aic}
\usage{
get_aic(clust_out)
}
\arguments{
\item{clust_out}{list of cluster dataframes for number of clusters}
}
\description{
Get the set of clusters which minimizes aic
}
\seealso{
Other cluster_properties: 
\code{\link{make_clust_col_name}()},
\code{\link{make_clust_col}()}

Other cluster_functions: 
\code{\link{cluster_kmeans_min}()},
\code{\link{cluster_kmeans}()}
}
\concept{cluster_functions}
\concept{cluster_properties}
