% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clust_col.R
\name{make_clust_col}
\alias{make_clust_col}
\title{Function for making the columns for clusters}
\usage{
make_clust_col(i, rows = character(), col_val = NA)
}
\arguments{
\item{i}{the cluster number}

\item{rows}{Optional input of the row names for the dataframe}

\item{col_val}{Default value for the column. default \:NA}
}
\description{
Function for making the columns for clusters
}
\seealso{
Other cluster_labels: 
\code{\link{make_clust_col_name}()}

Other cluster_properties: 
\code{\link{make_clust_col_name}()}
}
\concept{cluster_labels}
\concept{cluster_properties}
