% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clust_col_name.R
\name{make_clust_col_name}
\alias{make_clust_col_name}
\title{Function for making the label for a cluster column}
\usage{
make_clust_col_name(i)
}
\arguments{
\item{i}{the cluster number}
}
\description{
Function for making the label for a cluster column
}
\seealso{
Other cluster_labels: 
\code{\link{make_clust_col}()}

Other cluster_properties: 
\code{\link{make_clust_col}()}
}
\concept{cluster_labels}
\concept{cluster_properties}
