% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_by_end_col.R
\name{rescale_by_end_col}
\alias{rescale_by_end_col}
\title{Multiply all by the final terms in a row by the final term.
Repeat for all rows.}
\usage{
rescale_by_end_col(matin)
}
\arguments{
\item{matin}{matrix whose rows are to be rescaled by the final column}
}
\description{
Multiply all by the final terms in a row by the final term.
Repeat for all rows.
}
\seealso{
Other preconditioning_functions: 
\code{\link{convert_mat_to_angle_mat}()},
\code{\link{convert_point_to_angle}()},
\code{\link{crop_mat_colnames}()},
\code{\link{crop_mat_colnums}()},
\code{\link{crop_matrices}()},
\code{\link{reformat_data}()}
}
\concept{preconditioning_functions}
