% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cluster_kmeans.R, R/test_get_aic.R
\name{test_cluster_kmeans}
\alias{test_cluster_kmeans}
\title{Test the function for finding k clusters.}
\usage{
test_cluster_kmeans()

test_cluster_kmeans()
}
\description{
Test the function for finding k clusters.

Test the function for finding the aic of a set of clusters.
}
\details{
Checks:
* there are k clusters found
* The number of points with clusters assigned is the number of points input.

Checks:
* there result is a cluster dataframe with an aic column
* the aic is numeric
* the aic is the same on all columns
}
\seealso{
Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_join_matlist_on_name}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}

Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_join_matlist_on_name}()},
\code{\link{test_km_nan}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
