% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_find_aic.R
\name{test_find_aic}
\alias{test_find_aic}
\title{Test the function`find_all_ic`}
\usage{
test_find_aic()
}
\description{
Test the column names after using `find_all_ic`
on a dummy dataframe.
}
\seealso{
Other tests: 
\code{\link{test_add_nclust_cols_empty}()},
\code{\link{test_add_nclust_cols_onerow}()},
\code{\link{test_add_np_cols_empty}()},
\code{\link{test_add_np_cols_onerow}()},
\code{\link{test_calc_clust_dist_col}()},
\code{\link{test_calc_clust_dist}()},
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_find_low_percent_col_mat}()},
\code{\link{test_find_low_percent_col}()},
\code{\link{test_km_nan}()},
\code{\link{test_make_clust_col}()},
\code{\link{test_na_check}()}
}
\concept{tests}
