% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_km_nan.R
\name{test_km_nan}
\alias{test_km_nan}
\title{Test the function for finding k clusters.}
\usage{
test_km_nan()
}
\description{
Test the function for finding k clusters.
}
\details{
Checks:
* there are k clusters found
* The number of points with clusters assigned is the number of points input.
}
\seealso{
Other tests: 
\code{\link{test_calc_member_dist_cent}()},
\code{\link{test_cluster_kmeans_min}()},
\code{\link{test_cluster_kmeans}()},
\code{\link{test_convert_point_to_angle}()},
\code{\link{test_crop_matrices}()},
\code{\link{test_get_aic}()},
\code{\link{test_join_matlist_on_name}()},
\code{\link{test_reformat_data}()},
\code{\link{test_rescale_by_end_col}()}
}
\concept{tests}
